# backend/src/modules/verification/verification.controller.ts
import { Controller, Get, Post, Put, Param, UseGuards, Body, Req } from "@nestjs/common";
import { VerificationService } from "./verification.service";
import { JwtAuthGuard } from "../../shared/guards/jwt-auth.guard";
import { RolesGuard } from "../../shared/guards/roles.guard";
import { Roles } from "../../shared/decorators/roles.decorator";
import { PermissionsGuard } from "../../shared/guards/permissions.guard";
import { Permissions } from "../../shared/decorators/permissions.decorator";
import { Permission } from "../../shared/constants/permissions";
import { IsNotEmpty, IsString, MaxLength, IsOptional } from "class-validator";
import { ApiProperty } from "@nestjs/swagger";

export class RejectProviderDto {
  @ApiProperty({ description: "Rejection reason" })
  @IsNotEmpty()
  @IsString()
  @MaxLength(1000)
  reason: string;
}

export class RequestCorrectionsDto {
  @ApiProperty({ description: "Correction request comments/instructions" })
  @IsNotEmpty()
  @IsString()
  @MaxLength(1000)
  comments: string;
}

export class AssignReviewerDto {
  @ApiProperty({ description: "Admin Reviewer ID" })
  @IsNotEmpty()
  @IsString()
  reviewerId: string;
}

export class SanctionProviderDto {
  @ApiProperty({ description: "Sanction reason or compliance violation details" })
  @IsNotEmpty()
  @IsString()
  @MaxLength(1000)
  reason: string;

  @ApiProperty({ required: false, description: "Sanction type: suspended, revoked, probation" })
  @IsOptional()
  @IsString()
  sanctionType?: string;
}

@Controller("verification")
@UseGuards(JwtAuthGuard, RolesGuard, PermissionsGuard)
@Roles("admin", "super_admin", "verification_admin", "verifier")
export class VerificationController {
  constructor(private readonly verificationService: VerificationService) {}

  @Get(["", "pending"])
  @Permissions(Permission.CREDENTIALS_REVIEW)
  async getVerifications() {
    return this.verificationService.getVerificationQueue();
  }

  @Get("expiring-licenses")
  @Permissions(Permission.CREDENTIALS_REVIEW)
  async getExpiringLicenses() {
    return this.verificationService.getExpiringLicenses();
  }

  @Post(":id/assign-reviewer")
  @Permissions(Permission.CREDENTIALS_APPROVE)
  async assignReviewer(
    @Param("id") id: string,
    @Body() body: AssignReviewerDto,
    @Req() req: any
  ) {
    const actorId = req.user?.id || "u-admin";
    return this.verificationService.assignReviewer(id, body.reviewerId, actorId);
  }

  @Put(":id")
  @Permissions(Permission.CREDENTIALS_APPROVE)
  async updateVerificationDecision(
    @Param("id") id: string,
    @Body() body: any,
    @Req() req: any
  ) {
    const actorId = req.user?.id || req.user?.sub || "u-admin";
    if (body.status === "verified" || body.status === "approved") {
      return this.verificationService.approveProvider(id, actorId);
    } else if (body.status === "rejected") {
      return this.verificationService.rejectProvider(id, body.notes || "Application rejected", actorId);
    }
    return this.verificationService.requestCorrections(id, body.notes || "Corrections required", actorId);
  }

  @Post(":id/approve")
  @Permissions(Permission.CREDENTIALS_APPROVE)
  async approve(@Param("id") id: string, @Req() req: any) {
    const actorId = req.user?.id || "u-admin";
    return this.verificationService.approveProvider(id, actorId);
  }

  @Post(":id/reject")
  @Permissions(Permission.CREDENTIALS_APPROVE)
  async reject(
    @Param("id") id: string,
    @Body() body: RejectProviderDto,
    @Req() req: any
  ) {
    const actorId = req.user?.id || "u-admin";
    return this.verificationService.rejectProvider(id, body.reason, actorId);
  }

  @Post(":id/request-corrections")
  @Permissions(Permission.CREDENTIALS_APPROVE)
  async requestCorrections(
    @Param("id") id: string,
    @Body() body: RequestCorrectionsDto,
    @Req() req: any
  ) {
    const actorId = req.user?.id || "u-admin";
    return this.verificationService.requestCorrections(id, body.comments, actorId);
  }

  @Post(":id/sanction")
  @Permissions(Permission.CREDENTIALS_APPROVE)
  async sanction(
    @Param("id") id: string,
    @Body() body: SanctionProviderDto,
    @Req() req: any
  ) {
    const actorId = req.user?.id || "u-admin";
    return this.verificationService.sanctionProvider(id, body.reason, body.sanctionType || "suspended", actorId);
  }
}