# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/verification/{id}/assign-reviewer": {
      "post": {
        "operationId": "VerificationController_assignReviewer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignReviewerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/verification/{id}": {
      "put": {
        "operationId": "VerificationController_updateVerificationDecision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/verification/{id}/approve": {
      "post": {
        "operationId": "VerificationController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/verification/{id}/reject": {
      "post": {
        "operationId": "VerificationController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/verification/{id}/request-corrections": {
      "post": {
        "operationId": "VerificationController_requestCorrections",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCorrectionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/verification/{id}/sanction": {
      "post": {
        "operationId": "VerificationController_sanction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SanctionProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Merihcare API Specification",
    "description": "Comprehensive API contract documentation for Merihcare clinics, users, and administrative portals.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AssignReviewerDto": {
        "type": "object",
        "properties": {
          "reviewerId": {
            "type": "string",
            "description": "Admin Reviewer ID"
          }
        },
        "required": [
          "reviewerId"
        ]
      },
      "RejectProviderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Rejection reason"
          }
        },
        "required": [
          "reason"
        ]
      },
      "RequestCorrectionsDto": {
        "type": "object",
        "properties": {
          "comments": {
            "type": "string",
            "description": "Correction request comments/instructions"
          }
        },
        "required": [
          "comments"
        ]
      },
      "SanctionProviderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Sanction reason or compliance violation details"
          },
          "sanctionType": {
            "type": "string",
            "description": "Sanction type: suspended, revoked, probation"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}